#111: Moved queries related to mails to seperate files | edited commands to get correct queries
