+++ src/Intracto/SecretSantaBundle/Command/SendPoolStatusCommand.php
@@ -32,9 +32,9 @@ class SendPoolStatusCommand extends ContainerAwareCommand
         $container = $this->getContainer();
         /** @var EntityManager $em */
         $em = $container->get('doctrine')->getManager();
-        $entryQuery = $container->get('intracto_secret_santa.entry');
+        $entryMailQuery = $container->get('intracto_secret_santa.entry_mail');
         $mailerService = $container->get('intracto_secret_santa.mail');
-        $poolAdmins = $entryQuery->findAllAdminsForPoolStatusMail();
+        $poolAdmins = $entryMailQuery->findAllAdminsForPoolStatusMail();
         $timeNow = new \DateTime();
 
         foreach ($poolAdmins as $poolAdmin) {
