+++ src/Intracto/SecretSantaBundle/Command/SendWishlistUpdatedCommand.php
@@ -32,9 +32,9 @@ class SendWishlistUpdatedCommand extends ContainerAwareCommand
         $container = $this->getContainer();
         /** @var EntityManager $em */
         $em = $container->get('doctrine')->getManager();
-        $entryQuery = $container->get('intracto_secret_santa.entry');
+        $entryMailQuery = $container->get('intracto_secret_santa.entry_mail');
         $mailerService = $container->get('intracto_secret_santa.mail');
-        $secret_santas = $entryQuery->findAllToRemindOfUpdatedWishlist();
+        $secret_santas = $entryMailQuery->findAllToRemindOfUpdatedWishlist();
         $timeNow = new \DateTime();
 
         foreach ($secret_santas as $secret_santa) {
